Move some toolchain's classes
