+++ src/main/kotlin/org/rust/cargo/runconfig/CargoTestRunState.kt
@@ -14,12 +14,12 @@ import com.intellij.execution.runners.ExecutionEnvironment
 import com.intellij.execution.runners.ProgramRunner
 import com.intellij.execution.testframework.autotest.ToggleAutoTestAction
 import com.intellij.util.text.SemVer
+import org.rust.cargo.CargoCommandLine
 import org.rust.cargo.runconfig.buildtool.CargoPatch
 import org.rust.cargo.runconfig.command.CargoCommandConfiguration
 import org.rust.cargo.runconfig.console.CargoTestConsoleBuilder
-import org.rust.cargo.toolchain.CargoCommandLine
 import org.rust.cargo.toolchain.RustChannel
-import org.rust.cargo.toolchain.components.RustcVersion
+import org.rust.cargo.toolchain.impl.RustcVersion
 import java.time.LocalDate
 
 class CargoTestRunState(
