+++ src/main/kotlin/org/rust/cargo/runconfig/wasmpack/WasmPackCommandConfiguration.kt
@@ -12,6 +12,7 @@ import com.intellij.openapi.options.SettingsEditor
 import com.intellij.openapi.project.Project
 import com.intellij.util.execution.ParametersListUtil
 import org.jdom.Element
+import org.rust.cargo.WasmPackCommandLine
 import org.rust.cargo.project.model.cargoProjects
 import org.rust.cargo.project.settings.toolchain
 import org.rust.cargo.runconfig.command.workingDirectory
@@ -20,7 +21,6 @@ import org.rust.cargo.runconfig.readString
 import org.rust.cargo.runconfig.ui.WasmPackCommandConfigurationEditor
 import org.rust.cargo.runconfig.writePath
 import org.rust.cargo.runconfig.writeString
-import org.rust.cargo.toolchain.WasmPackCommandLine
 import java.nio.file.Path
 
 class WasmPackCommandConfiguration(
