+++ src/main/kotlin/org/rust/cargo/toolchain/RsToolchain.kt
@@ -15,8 +15,8 @@ import org.rust.cargo.toolchain.binaries.Grcov
 import org.rust.cargo.toolchain.binaries.WasmPack
 import org.rust.cargo.toolchain.components.Cargo
 import org.rust.cargo.toolchain.components.Rustc
-import org.rust.cargo.toolchain.components.RustcVersion
 import org.rust.cargo.toolchain.components.Rustfmt
+import org.rust.cargo.toolchain.impl.RustcVersion
 import java.io.File
 import java.nio.file.Files
 import java.nio.file.Path
