+++ src/main/kotlin/org/rust/cargo/toolchain/impl/BuildScriptsInfo.kt
@@ -6,7 +6,6 @@
 package org.rust.cargo.toolchain.impl
 
 import org.rust.cargo.project.workspace.PackageId
-import org.rust.cargo.toolchain.BuildScriptMessage
 
 class BuildScriptsInfo(private val messages: Map<PackageId, BuildScriptMessage>) {
     val containsOutDirInfo: Boolean get() = messages.entries.firstOrNull()?.value?.out_dir != null
